# Speedup sc.get.obs_df