--- a/scanpy/get.py
+++ b/scanpy/get.py
@@ -3,7 +3,7 @@ from typing import Optional, Iterable, Tuple
 
 import numpy as np
 import pandas as pd
-from scipy.sparse import spmatrix
+from scipy.sparse import spmatrix, issparse
 
 from anndata import AnnData
 
@@ -181,18 +181,17 @@ def obs_df(
             var_idx = adata.var_names.get_indexer(var_names)
 
         # for backed AnnData is important that the indices are ordered
-        var_order = np.argsort(var_idx)
-        matrix = X[:, var_idx[var_order]]
+        if adata.isbacked:
+            var_order = np.argsort(var_idx)
+            matrix = X[:, var_idx[var_order]][:, np.argsort(var_order)]
+        else:
+            matrix = X[:, var_idx]
 
         from scipy.sparse import issparse
 
         if issparse(matrix):
             matrix = matrix.toarray()
-        df = df.join(
-            pd.DataFrame(
-                matrix, columns=np.array(var_symbol)[var_order], index=adata.obs.index
-            )
-        )
+        df = df.join(pd.DataFrame(matrix, columns=var_symbol, index=adata.obs.index))
 
     # add obs values
     if len(obs_names) > 0:
@@ -264,18 +263,17 @@ def var_df(
         obs_idx = adata.obs_names.get_indexer(obs_names)
 
         # for backed AnnData is important that the indices are ordered
-        obs_order = np.argsort(obs_idx)
-        matrix = X[obs_idx[obs_order], :]
+        if adata.isbacked:
+            obs_order = np.argsort(obs_idx)
+            matrix = X[obs_idx[obs_order], :][np.argsort(obs_order)]
+        else:
+            matrix = X[obs_idx, :]
         from scipy.sparse import issparse
 
         if issparse(matrix):
             matrix = matrix.toarray()
 
-        df = df.join(
-            pd.DataFrame(
-                matrix.T, columns=np.array(obs_names)[obs_order], index=adata.var.index
-            )
-        )
+        df = df.join(pd.DataFrame(matrix.T, columns=obs_names, index=adata.var.index))
 
     # add obs values
     if len(var_names) > 0: