--- a/scanpy/tests/test_get.py
+++ b/scanpy/tests/test_get.py
@@ -19,77 +19,66 @@ def adata():
         var=pd.DataFrame(
             {"gene_symbols": ["genesymbol1", "genesymbol2"]}, index=["gene1", "gene2"]
         ),
-        layers={"double": np.ones((2, 2)) * 2},
+        layers={"double": np.ones((2, 2), dtype=int) * 2},
+        dtype=int,
     )
 
 
 def test_obs_df(adata):
-    adata.obsm["eye"] = np.eye(2)
-    adata.obsm["sparse"] = sparse.csr_matrix(np.eye(2))
+    adata.obsm["eye"] = np.eye(2, dtype=int)
+    adata.obsm["sparse"] = sparse.csr_matrix(np.eye(2), dtype='float64')
 
     adata.raw = AnnData(
         X=np.zeros((2, 2)),
         var=pd.DataFrame({"gene_symbols": ["raw1", "raw2"]}, index=["gene1", "gene2"]),
+        dtype='float64',
     )
-    assert np.all(
-        np.equal(
-            sc.get.obs_df(
-                adata, keys=["gene2", "obs1"], obsm_keys=[("eye", 0), ("sparse", 1)]
-            ),
-            pd.DataFrame(
-                {"gene2": [1, 1], "obs1": [0, 1], "eye-0": [1, 0], "sparse-1": [0, 1]},
-                index=adata.obs_names,
-            ),
-        )
+    pd.testing.assert_frame_equal(
+        sc.get.obs_df(
+            adata, keys=["gene2", "obs1"], obsm_keys=[("eye", 0), ("sparse", 1)]
+        ),
+        pd.DataFrame(
+            {"gene2": [1, 1], "obs1": [0, 1], "eye-0": [1, 0], "sparse-1": [0.0, 1.0]},
+            index=adata.obs_names,
+        ),
     )
-    assert np.all(
-        np.equal(
-            sc.get.obs_df(
-                adata,
-                keys=["genesymbol2", "obs1"],
-                obsm_keys=[("eye", 0), ("sparse", 1)],
-                gene_symbols="gene_symbols",
-            ),
-            pd.DataFrame(
-                {
-                    "genesymbol2": [1, 1],
-                    "obs1": [0, 1],
-                    "eye-0": [1, 0],
-                    "sparse-1": [0, 1],
-                },
-                index=adata.obs_names,
-            ),
-        )
+    pd.testing.assert_frame_equal(
+        sc.get.obs_df(
+            adata,
+            keys=["genesymbol2", "obs1"],
+            obsm_keys=[("eye", 0), ("sparse", 1)],
+            gene_symbols="gene_symbols",
+        ),
+        pd.DataFrame(
+            {
+                "genesymbol2": [1, 1],
+                "obs1": [0, 1],
+                "eye-0": [1, 0],
+                "sparse-1": [0.0, 1.0],
+            },
+            index=adata.obs_names,
+        ),
     )
-    assert np.all(
-        np.equal(
-            sc.get.obs_df(adata, keys=["gene2", "obs1"], layer="double"),
-            pd.DataFrame({"gene2": [2, 2], "obs1": [0, 1]}, index=adata.obs_names),
-        )
+    pd.testing.assert_frame_equal(
+        sc.get.obs_df(adata, keys=["gene2", "obs1"], layer="double"),
+        pd.DataFrame({"gene2": [2, 2], "obs1": [0, 1]}, index=adata.obs_names),
     )
-    assert np.all(
-        np.equal(
-            sc.get.obs_df(
-                adata, keys=["raw2", "obs1"], gene_symbols="gene_symbols", use_raw=True
-            ),
-            pd.DataFrame({"raw2": [0, 0], "obs1": [0, 1]}, index=adata.obs_names),
-        )
+
+    pd.testing.assert_frame_equal(
+        sc.get.obs_df(
+            adata, keys=["raw2", "obs1"], gene_symbols="gene_symbols", use_raw=True
+        ),
+        pd.DataFrame({"raw2": [0.0, 0.0], "obs1": [0, 1]}, index=adata.obs_names),
     )
     # test only obs
-    assert np.all(
-        np.equal(
-            sc.get.obs_df(adata, keys=["obs1", "obs2"]),
-            pd.DataFrame(
-                {"obs1": [0, 1], "obs2": ["a", "b"]}, index=["cell1", "cell2"]
-            ),
-        )
+    pd.testing.assert_frame_equal(
+        sc.get.obs_df(adata, keys=["obs1", "obs2"]),
+        pd.DataFrame({"obs1": [0, 1], "obs2": ["a", "b"]}, index=["cell1", "cell2"]),
     )
     # test only var
-    assert np.all(
-        np.equal(
-            sc.get.obs_df(adata, keys=["gene1", "gene2"]),
-            pd.DataFrame({"gene1": [1, 1], "gene2": [1, 1]}, index=adata.obs_names),
-        )
+    pd.testing.assert_frame_equal(
+        sc.get.obs_df(adata, keys=["gene1", "gene2"]),
+        pd.DataFrame({"gene1": [1, 1], "gene2": [1, 1]}, index=adata.obs_names),
     )
     badkeys = ["badkey1", "badkey2"]
     with pytest.raises(KeyError) as badkey_err: