By using array slicing, this codes improves ~10 fold the speed of `sc.get.obs_df()`.

```PYTHON
%timeit sc.get.obs_df(adata, list(adata.var_names[:100]) + ['louvain'])
````
before:
`40.6 ms ± 2.38 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)`

after:
`4.45 ms ± 228 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`