@fidelram 2020-11-20
I like the ideas of using `adata.raw` when raw data wants to be used in general.  This is an elegant solution to a source of endless confusion.

Regarding `var_df` I can remove the `use_raw`. However, I consider that since this option is everywhere it should be here as well. I find odd that is not present.  I don't see a problem that the size of the dataframe will be different that would be expected.