@ivirshup 2020-11-20
It's not just that the length is different, is that `sc.get.obs_df(adata, ["col"], use_raw=x)["col"]` is the same regardless of the value of `x`, but it's different for `var_df`. I think it's easier to build code around functions with more orthogonal arguments.

> However, I consider that since this option is everywhere it should be here as well. 

Could we add an example of `sc.get.var_df(adata.raw, ...)`, leave out `use_raw` for now, and see if anyone complains?

I've been trying to leave out `use_raw` on functions where variable length matters anyways. For example: `adata.var_vector`.