@fidelram 2020-11-25
The previous test failed but is not clear to me why, as it passes the local tests (anndata 0.7.5). It seems that on travis server,  backed slicing requires integer indices and will not work with a boolean vector. I changed to sorted integers hoping that this will solve the issue.