@ivirshup 2020-12-03
My thinking on my change is that I would like all the code that handles backed mode to be cleanly separated. I think this should be handled more cleanly on the `anndata` side, and once that's been done it's easier to replace the backed mode specific code if it's all together.